First completed rough draft of hangman.
My goal for this project was to create a basic hangman game that will present to the user the hangman target word in blanks and if the user guessed a letter correctly, that letter will be shown, but the rest of the letters will remain blank until the user guessed those letters correctly. The code in this file is a continuation from the first file that I uploaded in the 'hang_game' repository (hangman_edit_3).